chat-backend/pyproject.toml
[project]
name = "chat-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Vicente Mattos",email = "vicentemattosf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.116.2,<0.117.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "starlette (>=0.48.0,<0.49.0)",
    "groq (>=0.32.0,<0.33.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.13.1,<0.14.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint ='ruff check'
pre_format = 'ruff check --fix'
format ='ruff format'
run ='fastapi dev chat_backend/app.py'
pre_test = 'task lint'
test ='pytest -s -x --cov=chat_backend -vv'
post_test = "coverage html"

[tool.pylint.main]
extension-pkg-allow-list = ["pydantic"]